refact: move naver response to naver package
